refactor - rename provider
reflect specialized case in the name of the class implementing the interface, this might not be necessary now but nice to have if adding more exchange providers in the future, also moved it back outside of models
